refactor(recoil): fonte unica (recoil.js); editor usa o RECUO REAL do jogo
O recuo do editor era um sim inventado (subia reto) e divergia do jogo.
Agora:
- public/js/recoil.js: FONTE UNICA (buildRecoilPattern, RECOIL_PARAMS/PATTERN/
  CLASS, REC_DEG, REC, vmKickAmp) — extraida do game.js;
- game.js importa dela (nenhuma logica de recuo reescrita);
- dev.html (aba Viewmodel) usa o MESMO modelo: view-punch que serpenteia
  (esq/dir), aleatorio (+-30%), segura na rajada e recupera (hold/tau/perm),
  + kick da arma pela vmKickAmp real. A secao 'Recuo (real)' edita recoil.js
  ao vivo (forca/padrao/timing), igual a aba do jogo.
Agora o recuo do editor == recuo do jogo (mesma fonte).
